Fixed issues with brightness and blinking. But low brightness causes issues with tubes now

+++ components/nixie_lib/src/nixie_tube_base.cpp
@@ -216,5 +216,5 @@ uint8_t NixieTubeBase::get_actual_brightness()
     {
         brightness = m_fake_brightness;
     }
-    return brightness * (uint16_t)m_user_brightness_fraction / 100;
+    return (uint16_t)brightness * (uint16_t)m_user_brightness_fraction / 100;
 }
